refactor: extract AnalyzerResultsAcceptService from controller
Constitution Principle IV: business logic belongs in services, not controllers.

Extract ~700 lines of accept/save logic from AnalyzerResultsController
(1563→875 lines) into AnalyzerResultsAcceptServiceImpl:
- acceptAndPersist(): orchestrates the full accept flow
- buildSampleGroupings(): creates Sample/SampleItem/Analysis/Result
- 4 entry-state path handlers (noEntry, demographics, sample+demo, prevAnalyzer)
- All supporting methods (result creation, analysis population, reflexes)

SampleGrouping extracted from controller inner class to standalone valueholder.

Reverts the 3 broken guards that silently skipped record creation:
- getOrCreateSampleItem: restore .get(0) without isEmpty guard
- createGroupForSampleAndDemographicsEntered: restore sampleItem == null check
- getTypeOfSampleId: restore .get(0) without isEmpty guard
The TypeOfSampleTestServiceImpl routing fix (DAO HQL method) is the real fix.

Also: synchronous bridge registration (registerWithBridge returns boolean,
bridgeRegistered in 201 response).

54/54 tests pass.
